+++ changelog.md
@@ -9,6 +9,10 @@
 - `29/7/2026`: [LuauAPI has an example/template mod now](https://github.com/ItMe12s/luauapi-example-mod).
 - `11/7/2026`: Bindings that are automatically picked up likely won't be included in the changelog.
 
+## v0.1.0-beta.27
+
+add stuff here...
+
 ## v0.1.0-beta.26
 
 - Redesigned `gd3d` internals for code quality, performance and iOS support.